// EMD.Client/src/app/Models/employee-models/employee.model.ts
export interface EmployeeModel {
  employeeId?: number;

  employeeName: string;

  phoneNo: string;

  email: string;

  password: string;

  role: string;

  city: string;

  state: string;

  pincode: string;

  address: string;

  altPhoneNo?: string;

  designationId: number;
}